Se agrega parámetro de vehicle repository y se obtene el cilindraje que retorna al momento de crear el objeto según su tipo

+++ domain/src/main/java/com/ceiba/domain/repository/VehicleEnterRepository.kt
@@ -3,5 +3,5 @@ package com.ceiba.domain.repository
 import com.ceiba.domain.aggregate.ParkingEntranceExit
 
 interface VehicleEnterRepository {
-    suspend fun enterVehicle(parkingEntranceExit: ParkingEntranceExit): Long?
+    suspend fun enterVehicle(parkingEntranceExit: ParkingEntranceExit): Int?
 }
